requirements/dev.txt
git+https://github.com/timoniq/vkbottle.git
SQLAlchemy==1.4.0b1
asyncpg
alembic
python-dotenv
flake8
pytest